pyproject.toml
[project]
name = "support-swarm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "jinja2>=3.1.6",
    "langchain>=1.2.12",
    "langchain-anthropic>=1.3.4",
    "langchain-openai>=1.1.11",
    "langgraph>=1.1.1",
    "pgvector>=0.4.2",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.48",
]
